QOL + Interactables
Made some minor QOL changes in the form of setting a bunch of lights to static for minor optimization gains.

Edited the WorldItem script to be of dual purpose, if it has an item assigned, it serves as a pickup, otherwise, it serves as a piece of the environment that can be interacted with for flavour text.